refactor(transactions): extract list-rendering layer to sibling file (#272, #268, #267)
Move the list-rendering computed properties (`listView`,
`transactionsList`, `emptyStateOverlay`) and the file-scope
`TransactionListCSVImportAddons` ViewModifier from
`TransactionListView.swift` into
`TransactionListView+List.swift`. Drops the original from 553 → 319
lines (under the 400 file_length limit).

Why this seam: the inspector is already extracted into
`OptionalTransactionInspector`, and `body`'s sheets/alerts are too
tightly fused to the modifier chain to extract. The list-rendering
block is the only ~210-line cohesive slice that moves cleanly.

Side fixes that fell out of the move (each clears a baseline entry
that would otherwise re-key under the new file path):

- Extract `transactionRow(for:)`, `rowContextMenu(for:)`, and
  `loadMoreFooter` helpers from the inline `List`/`ForEach`
  closure — both closure bodies were over the 30-line limit.
- Rename `tx` → `transaction` in `CreateRuleFromTransactionSheet`
  closure to clear the `identifier_name` violation.

Access-level widenings (forced by `strict_fileprivate` — the rule
disallows `fileprivate` on extensions, so `internal` is the
smallest legal scope when a private helper is referenced from a
sibling-file extension):

- `@State private` → `@State` for state mutated/bound by the
  extension's views (`activeFilter`, `showFilterSheet`,
  `positionsInput`, `positionsRange`, `searchText`,
  `transactionPendingDelete`, `createRuleFromTransaction`).
- `private var` → `var` for read-from-extension properties
  (`filter`, `displayTitle`, `selectedTransaction`,
  `selectedTransactionBinding`, `filteredTransactions`).
- `private func createNewTransaction()` → `func` for the toolbar
  Add button.
- `TransactionListCSVImportAddons` was file-scope private; now
  internal (definition moved to the sibling file).

State that stayed `private` (only used inside the original file):
`_externalSelection`, `_internalSelection`, `handlesOwnInspector`,
`showError`, `errorMessage`, `searchFieldFocused`,
`ingestDroppedURLs`. Each widening site has a short comment
explaining why.

No behavioural change.
